+++ concordium-node/build.rs
@@ -673,6 +673,17 @@ fn build_grpc2(proto_root_input: &str) -> std::io::Result<()> {
                 .server_streaming()
                 .build(),
         )
+        .method(
+            tonic_build::manual::Method::builder()
+                .name("dry_run")
+                .route_name("DryRun")
+                .input_type("crate::grpc2::types::DryRunRequest")
+                .client_streaming()
+                .output_type("Vec<u8>")
+                .codec_path("crate::grpc2::RawCodec")
+                .server_streaming()
+                .build(),
+        )
         .build();
     // Due to the slightly hacky nature of the RawCodec (i.e., it does not support
     // deserialization) we cannot build the client. But we also don't need it in the
@@ -689,6 +700,18 @@ fn build_grpc2(proto_root_input: &str) -> std::io::Result<()> {
             .build_client(false)
             .file_descriptor_set_path(descriptor_path)
             .compile(&[&health], &[proto_root_input])
+            .expect("Failed to compile gRPC health definitions!");
+    }
+    {
+        let grpc_health_v1 = format!("{}/grpc/health/v1/health.proto", proto_root_input);
+        let descriptor_path = std::path::PathBuf::from(env::var("OUT_DIR").unwrap())
+            .join("grpc_health_v1_descriptor.bin");
+        // build the health service with reflection support
+        tonic_build::configure()
+            .build_server(true)
+            .build_client(false)
+            .file_descriptor_set_path(descriptor_path)
+            .compile(&[&grpc_health_v1], &[proto_root_input])
             .expect("Failed to compile gRPC health definitions!");
     }
     Ok(())