+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+- Remove V1 GRPC API from the node. This removes configuration options
+  `CONCORDIUM_NODE_RPC_SERVER_PORT`, `CONCORDIUM_NODE_RPC_SERVER_ADDRESS`, 
+  `CONCORDIUM_NODE_RPC_SERVER_TOKEN`.
+
 ## 6.2.0
 
 - Add an additional health-check service to the V2 GRPC API.