+++ CHANGELOG.md
@@ -4,9 +4,15 @@
 - The account map is now kept solely on disk in a separate lmdb database and it is no longer part of the internal block state database.
   This change results in less memory usage per account and a decrease in the growth of the database.
 
+## 6.2.0
+
+- Add an additional health-check service to the V2 GRPC API.
+  This service conforms to the [standard GRPC health service API](https://github.com/grpc/grpc-proto/blob/master/grpc/health/v1/health.proto).
+- New `DryRun` endpoint that allows simulating the execution of transactions.
+
 ## 6.1.7
 
- - Add load-shedding to the V2 GRPC API. In particular, if at the time of the
+- Add load-shedding to the V2 GRPC API. In particular, if at the time of the
   request the node is already handling more than
   `CONCORDIUM_NODE_GRPC2_MAX_CONCURRENT_REQUESTS` requests then the incoming
   request will be immediately rejected.